Update: 2/5/18
I added something to the single quote part of the tokenizer and tried to see what it was doing by putting in systerm.out.println statements. There is a forever loop going on in the ending with single quote while loop. I will fix that. This is what I got so far.